add coins list
fetch coins data from coingecko API and send them to the client. On the client side add redux package to keep fetched data to a state in order to prevent multiple requests while browsing. Make app header fixed to always be visible in the coins page and render coins data